Various VFS-related bugfixes

+++ kernel/include/kernel/vfs/vfs.h
@@ -59,7 +59,7 @@ struct dirent;
 typedef struct {
     struct vfs_node* node; // Node pointed by this handle
     uint32_t id; // Handle ID (used internally by VFS)
-    uint32_t offset; // Current offset in bytes from the start of file
+    uint64_t offset; // Current offset in bytes from the start of file
     bool readable; // Is file readable?
     bool writeable; // Is file writeable?
     void* metadata; // Available to use by the filesystem
